Refactor Request type into adlu_parse.
This is both because it makes sense (so requests can be parsed directly in the module) and because it should allow for optimizing compile times by boxing some of the filters.
